Clean up jdt2famix.cmd
Dumping variable contents to stdout without any context is bad behavior.
In addition, these variables are not of interest to the jdt2famix user,
therefore they are no longer shown.

Having the unbalanced quotes one at the beginning of "set CLASSPATH" and
one on the last instance of "set CLASSPATH" was very confusing.