Computing the Cyclomatic Complexity metric.
The current available implementation of jdt2famix does not calculate
the metric cyclomatic complexity. In this patch we fix this issue, implementing
the algorithm detailed in

Empirical analysis of the relationship between CC and SLOC in a large corpus of Java methods and C functions.
D Landman, A Serebrenik, E Bouwers, JJ Vinju
Journal of Software: Evolution and Process 28 (7), 589-618